+++ crates/libwebrtc-sys/src/cpp/sys_audio_capture/win_capture.cc
@@ -1,3 +1,5 @@
+#if defined(WEBRTC_WIN)
+
 #include "rtc_base/logging.h"
 
 #include "libwebrtc-sys/include/sys_audio_capture/win_capture.h"
@@ -184,11 +186,10 @@ bool SysAudioSource::ProcessRecordedPart(bool firstInCycle) {
 
   if (recorded_samples_.size() >= kRecordingPart) {
     source_->OnData(recorded_samples_.data(), kBitsPerSample,
-                kRecordingFrequency, kRecordingChannels, kRecordingPart);
+                    kRecordingFrequency, kRecordingChannels, kRecordingPart);
 
-    recorded_samples_.erase(
-      recorded_samples_.begin(),
-      recorded_samples_.begin() + kRecordingPart);
+    recorded_samples_.erase(recorded_samples_.begin(),
+                            recorded_samples_.begin() + kRecordingPart);
 
     return true;
   }
@@ -230,43 +231,21 @@ bool SysAudioSource::ProcessRecordedPart(bool firstInCycle) {
                         << "AUDCLNT_BUFFERFLAGS_TIMESTAMP_ERROR.";
   }
 
-  const int16_t* res_data = nullptr;
   if (flags & AUDCLNT_BUFFERFLAGS_SILENT) {
     // Fill the provided number of frames in the vec with silence.
     recorded_samples_.insert(recorded_samples_.end(), numFramesAvailable, 0);
-    res_data = recorded_samples_.data();
   } else {
     const auto audioData = reinterpret_cast<const int16_t*>(buffer);
 
     if (recorded_samples_.empty() && numFramesAvailable == kRecordingPart) {
       // Skip intermediate buffer and write directly from the GetBuffer
       // result.
-      res_data = audioData;
+      source_->OnData(audioData, kBitsPerSample, kRecordingFrequency,
+                      kRecordingChannels, kRecordingPart);
     } else {
-      recorded_samples_.insert(
-          recorded_samples_.end(), audioData,
-          audioData + numFramesAvailable);
-
-      if (recorded_samples_.size() < kRecordingPart) {
-        static_cast<void>(
-          audio_client_activation_handler_->capture_client->ReleaseBuffer(
-            numFramesAvailable));
-
-        // Not enough data for 10 milliseconds.
-        return false;
-      }
-
-      res_data = recorded_samples_.data();
+      recorded_samples_.insert(recorded_samples_.end(), audioData,
+                               audioData + numFramesAvailable);
     }
-  }
-
-  source_->OnData(res_data, kBitsPerSample, kRecordingFrequency,
-                  kRecordingChannels, kRecordingPart);
-
-  if (!recorded_samples_.empty()) {
-    recorded_samples_.erase(
-      recorded_samples_.begin(),
-      recorded_samples_.begin() + kRecordingPart);
   }
 
   hr = audio_client_activation_handler_->capture_client->ReleaseBuffer(
@@ -284,3 +263,5 @@ bool SysAudioSource::ProcessRecordedPart(bool firstInCycle) {
 webrtc::scoped_refptr<bridge::LocalAudioSource> SysAudioSource::GetSource() {
   return source_;
 }
+
+#endif // WEBRTC_WIN