+++ crates/native/src/frb_generated.rs
@@ -954,7 +954,9 @@ impl SseDecode for crate::api::media::constraints::audio::AudioConstraints {
         deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer,
     ) -> Self {
         let mut var_deviceId = <Option<String>>::sse_decode(deserializer);
-        let mut var_processing = <crate::api::media::constraints::audio::AudioProcessingConstraints>::sse_decode(deserializer);
+        let mut var_processing = <Option<
+            crate::api::media::constraints::audio::AudioProcessingConstraints,
+        >>::sse_decode(deserializer);
         let mut var_isDisplay = <bool>::sse_decode(deserializer);
         return crate::api::media::constraints::audio::AudioConstraints {
             device_id: var_deviceId,
@@ -1604,6 +1606,23 @@ impl SseDecode
     ) -> Self {
         if (<bool>::sse_decode(deserializer)) {
             return Some(<crate::api::media::constraints::audio::AudioConstraints>::sse_decode(deserializer));
+        } else {
+            return None;
+        }
+    }
+}
+
+impl SseDecode
+    for Option<
+        crate::api::media::constraints::audio::AudioProcessingConstraints,
+    >
+{
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_decode(
+        deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer,
+    ) -> Self {
+        if (<bool>::sse_decode(deserializer)) {
+            return Some(<crate::api::media::constraints::audio::AudioProcessingConstraints>::sse_decode(deserializer));
         } else {
             return None;
         }
@@ -4337,7 +4356,9 @@ impl SseEncode for crate::api::media::constraints::audio::AudioConstraints {
         serializer: &mut flutter_rust_bridge::for_generated::SseSerializer,
     ) {
         <Option<String>>::sse_encode(self.device_id, serializer);
-        <crate::api::media::constraints::audio::AudioProcessingConstraints>::sse_encode(self.processing, serializer);
+        <Option<
+            crate::api::media::constraints::audio::AudioProcessingConstraints,
+        >>::sse_encode(self.processing, serializer);
         <bool>::sse_encode(self.is_display, serializer);
     }
 }
@@ -4915,6 +4936,23 @@ impl SseEncode
         <bool>::sse_encode(self.is_some(), serializer);
         if let Some(value) = self {
             <crate::api::media::constraints::audio::AudioConstraints>::sse_encode(value, serializer);
+        }
+    }
+}
+
+impl SseEncode
+    for Option<
+        crate::api::media::constraints::audio::AudioProcessingConstraints,
+    >
+{
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_encode(
+        self,
+        serializer: &mut flutter_rust_bridge::for_generated::SseSerializer,
+    ) {
+        <bool>::sse_encode(self.is_some(), serializer);
+        if let Some(value) = self {
+            <crate::api::media::constraints::audio::AudioProcessingConstraints>::sse_encode(value, serializer);
         }
     }
 }